Move save_shipping_rate_data to separate trait
